text_embedder/requirements.txt
flask
FlagEmbedding
python-dotenv